Update shader pipeline and message handling types
- Changed Window::ProcessMessages return type to std::optional<WPARAM> for better Windows API alignment
- Explicitly cast exit code in App::Go() to int
- Integrated compiled vertex shader (VertexShader.cso) in DrawTestTriangle and linked D3DCompiler
- Added VertexShader.hlsl and updated project files to include shader compilation and filtering
- Introduced InfoException in Graphics for enhanced error info
- Fixed minor issues in WinMain.cpp and utils.h

+++ RealTime3DRendering/App.cpp
@@ -13,7 +13,7 @@ int App::Go()
 	{
 		if (const auto exitCode = wnd.ProcessMessages())
 		{
-			return *exitCode;
+			return static_cast<int>(*exitCode);
 		}
 		DoFrame();
 	}
